+++ EDDI/EDDI.cs
@@ -4,6 +4,7 @@ using EddiDataProviderService;
 using EddiEvents;
 using EddiSpeechService;
 using EddiStarMapService;
+using Exceptionless;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using System;
@@ -17,6 +18,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Windows;
 using Utilities;
 
 namespace Eddi
@@ -93,6 +95,10 @@ namespace Eddi
             try
             {
                 Logging.Info(Constants.EDDI_NAME + " " + Constants.EDDI_VERSION + " starting");
+
+                // Exception handling
+                ExceptionlessClient.Default.Startup("vJW9HtWB2NHiQb7AwVQsBQM6hjWN1sKzHf5PCpW1");
+                ExceptionlessClient.Default.Configuration.SetVersion(Constants.EDDI_VERSION);
 
                 // Start by ensuring that our primary data structures have something in them.  This allows them to be updated
                 // from any source
@@ -290,6 +296,8 @@ namespace Eddi
             }
 
             Logging.Info(Constants.EDDI_NAME + " " + Constants.EDDI_VERSION + " stopped");
+
+            ExceptionlessClient.Default.Shutdown();
 
             started = false;
         }