+++ ChangeLog.md
@@ -5,6 +5,7 @@
   * Add system main star stellar class and age to VoiceAttack variables
   * Fix commander progress "trade" rating
   * Disable "distort voice on damage" effect until we can find a better distortion process
+  * Add separate exception logging system
 
 #2.0.11
   * Further fixes for renamed and missing fields in 2.2.02